fix(ci): unbreak #710 — split settings.rs, restore the arm its merge dropped (#717)
`fmt + clippy + test` failed in 7s on `feat/tool-switches-on-709`. As on main,
`check-file-size.sh` runs before the Rust toolchain installs, so the fast fail
hid a genuine compile error behind it. Three fixes:

1. **`stella-cli/src/settings.rs` crossed the 1500-line ratchet at 1623 lines.**
   It is not grandfathered, and the guard hard-blocks that case while explicitly
   forbidding a baseline entry — "the baseline only covers files that predate the
   guard, and this is the rule that stops the tree getting worse." So it is split,
   not exempted: the 868-line `#[cfg(test)] mod tests` moves to
   `stella-cli/src/settings/tests.rs`. That directory already holds seven sibling
   modules, and the same split has precedent in e72b02f0 (deck_ui.rs) and in
   stella-store's `*_tests.rs` files. settings.rs drops to 756 lines, the new file
   is 874 — neither needs an entry.

   Pure move, verified as such: the extracted body is byte-identical to the old
   inline body modulo one dedent level, all 27 `#[test]`s are still present, and
   `cargo fmt --check` passes on the result untouched. `use super::*` resolves to
   `crate::settings` from a file submodule exactly as it did inline, so nothing
   about name resolution changes.

2. **`StorageCmd::Prune(_)` was missing from `run_storage`'s match** —
   `error[E0004]: non-exhaustive patterns`. The merge commit 6e612078 kept the
   variant and its early-return dispatch but dropped the `unreachable!` arm.
   `origin/main` still has that arm (main.rs:1305) and main is green, so this was
   introduced by this branch's conflict resolution, not inherited. Restored.

3. **Four grandfathered files grew** — agent_tests.rs +157, agent.rs +32,
   candidate_ws.rs +24, command_deck.rs +1 — so the baseline is regenerated per
   the script's own instruction. It also *tightened* five ceilings where this
   branch shrank files (registry.rs 3052→3014, main.rs, stella-store/lib.rs,
   media.rs, and one more), so the ratchet moves net-inward. Still 32 entries.

   Note the ordering: regenerating the baseline before fix (2) records a ceiling
   that fix (2) then exceeds by the two lines it adds. Baseline last.

Gate, full workspace: check-no-scratch, check-action-pins, check-file-size,
`cargo fmt --check`, `cargo clippy --workspace --all-targets -- -D warnings`,
`RUSTDOCFLAGS=-D warnings cargo doc --workspace --no-deps`, and
`cargo test --workspace` — 3,559 tests across 61 suites, 0 failures.
