+++ scripts/file-size-baseline.txt
@@ -7,12 +7,12 @@
 # must be split instead. Raising an existing ceiling is allowed but is
 # never silent — it lands as a visible diff here, to be justified in
 # review like any other change.
-1715 stella-cli/src/agent_tests.rs
-2191 stella-cli/src/agent.rs
-1599 stella-cli/src/candidate_ws.rs
-4465 stella-cli/src/command_deck.rs
+1872 stella-cli/src/agent_tests.rs
+2223 stella-cli/src/agent.rs
+1623 stella-cli/src/candidate_ws.rs
+4466 stella-cli/src/command_deck.rs
 1503 stella-cli/src/contextgraph.rs
-1789 stella-cli/src/main.rs
+1776 stella-cli/src/main.rs
 1641 stella-cli/src/memory.rs
 2224 stella-context/src/store.rs
 2095 stella-core/src/bus.rs
@@ -28,11 +28,11 @@
 2504 stella-pipeline/src/pipeline.rs
 2019 stella-pipeline/src/pipeline/tests.rs
 2636 stella-protocol/src/event.rs
-2270 stella-store/src/lib.rs
+2268 stella-store/src/lib.rs
 2168 stella-store/src/tests.rs
 1884 stella-store/src/usage.rs
-1558 stella-tools/src/media.rs
-3052 stella-tools/src/registry.rs
+1556 stella-tools/src/media.rs
+3014 stella-tools/src/registry.rs
 1797 stella-tools/src/scripts.rs
 1670 stella-tui/src/deck_render.rs
 3940 stella-tui/src/deck_ui.rs
